requirements.txt
aiohttp==3.11.10
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==5.1.4
djangorestframework==3.15.2
playwright==1.49.1
ratelimit==2.2.1
django-filter==24.3
django-cors-headers==4.6.0
lxml==5.3.0
gunicorn==21.2.0
uvicorn[standard]==0.24.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0